Subject line (try to keep under 50 characters) Provide intial structure of index html file
This is the very first commit of the project it will include the
base structure of the HTML file, this is important because it is
the intial/essiential building block of a HTML code. Also the
updated README file explaining what the purpose and goals of this
project will be.

[Ticket: 1]